Swapped m2::Lab5 with m2::Lab6. Changed tasks in Lab5.
