+++ src/lab_m2/lab5/lab5.cpp
@@ -3,11 +3,15 @@
 #include <vector>
 #include <iostream>
 
-#include "stb/stb_image.h"
-
 using namespace std;
 using namespace m2;
 
+
+//Generates a random value between 0 and 1.
+inline float Rand01()
+{
+    return rand() / static_cast<float>(RAND_MAX);
+}
 
 /*
  *  To find out more about `FrameStart`, `Update`, `FrameEnd`
@@ -27,52 +31,74 @@ Lab5::~Lab5()
 
 void Lab5::Init()
 {
+    outputType = 0;
+
     auto camera = GetSceneCamera();
-    camera->SetPositionAndRotation(glm::vec3(0, 2, 4), glm::quat(glm::vec3(-30 * TO_RADIANS, 0, 0)));
+    camera->SetPositionAndRotation(glm::vec3(0, 2, 3.5), glm::quat(glm::vec3(-20 * TO_RADIANS, 0, 0)));
     camera->Update();
 
-    std::string texturePath = PATH_JOIN(window->props.selfDir, RESOURCE_PATH::TEXTURES, "cube");
-    std::string shaderPath = PATH_JOIN(window->props.selfDir, SOURCE_PATH::M2, "lab5", "shaders");
+    TextureManager::LoadTexture(PATH_JOIN(window->props.selfDir, RESOURCE_PATH::TEXTURES), "ground.jpg");
 
+    // Load a mesh from file into GPU memory
     {
-        Mesh* mesh = new Mesh("bunny");
-        mesh->LoadMesh(PATH_JOIN(window->props.selfDir, RESOURCE_PATH::MODELS, "animals"), "bunny.obj");
-        mesh->UseMaterials(false);
-        meshes[mesh->GetMeshID()] = mesh;
-    }
-
-    {
-        Mesh* mesh = new Mesh("cube");
+        Mesh* mesh = new Mesh("box");
         mesh->LoadMesh(PATH_JOIN(window->props.selfDir, RESOURCE_PATH::MODELS, "primitives"), "box.obj");
+        meshes[mesh->GetMeshID()] = mesh;
+    }
+
+    {
+        Mesh* mesh = new Mesh("plane");
+        mesh->LoadMesh(PATH_JOIN(window->props.selfDir, RESOURCE_PATH::MODELS, "primitives"), "plane50.obj");
         mesh->UseMaterials(false);
         meshes[mesh->GetMeshID()] = mesh;
     }
 
-    // Create a shader program for rendering to texture
+    // Load a mesh from file into GPU memory
     {
-        Shader *shader = new Shader("CubeMap");
-        shader->AddShader(PATH_JOIN(shaderPath, "CubeMap.VS.glsl"), GL_VERTEX_SHADER);
-        shader->AddShader(PATH_JOIN(shaderPath, "CubeMap.FS.glsl"), GL_FRAGMENT_SHADER);
-        shader->CreateAndLink();
-        shaders[shader->GetName()] = shader;
+        Mesh* mesh = new Mesh("sphere");
+        mesh->LoadMesh(PATH_JOIN(window->props.selfDir, RESOURCE_PATH::MODELS, "primitives"), "sphere.obj");
+        mesh->UseMaterials(false);
+        meshes[mesh->GetMeshID()] = mesh;
     }
 
-    // Create a shader program for rendering to texture
     {
-        Shader *shader = new Shader("ShaderNormal");
-        shader->AddShader(PATH_JOIN(shaderPath, "Normal.VS.glsl"), GL_VERTEX_SHADER);
-        shader->AddShader(PATH_JOIN(shaderPath, "Normal.FS.glsl"), GL_FRAGMENT_SHADER);
-        shader->CreateAndLink();
-        shaders[shader->GetName()] = shader;
+        Mesh* mesh = new Mesh("quad");
+        mesh->LoadMesh(PATH_JOIN(window->props.selfDir, RESOURCE_PATH::MODELS, "primitives"), "quad.obj");
+        mesh->UseMaterials(false);
+        meshes[mesh->GetMeshID()] = mesh;
     }
 
-    cubeMapTextureID = UploadCubeMapTexture(
-        PATH_JOIN(texturePath, "pos_x.png"),
-        PATH_JOIN(texturePath, "pos_y.png"),
-        PATH_JOIN(texturePath, "pos_z.png"),
-        PATH_JOIN(texturePath, "neg_x.png"),
-        PATH_JOIN(texturePath, "neg_y.png"),
-        PATH_JOIN(texturePath, "neg_z.png"));
+    LoadShader("Render2Texture");
+    LoadShader("Composition");
+    LoadShader("LightPass");
+
+    auto resolution = window->GetResolution();
+
+    frameBuffer = new FrameBuffer();
+    frameBuffer->Generate(resolution.x, resolution.y, 3);
+    //frameBuffer contains 3 textures (position, normal and color)
+
+    lightBuffer = new FrameBuffer();
+    lightBuffer->Generate(resolution.x, resolution.y, 1, false);
+    //lightBuffer contains 1 texture (light accumulation)
+
+    for (int i = 0; i < 40; ++i)
+    {
+        LightInfo lightInfo;
+
+        // TODO(student): Set lightInfo with random position, random color
+        // and a random radius for each light source.
+        // You can use the Rand01 function defined above.
+        // The chosen position is between (-10, 0, -10) and (10, 3, 10)
+        // The chosen color is between (0, 0, 0) and (1, 1, 1).
+        // The chosen radius is between 3 and 4.
+
+        lightInfo.position = glm::vec3(0.0f);
+        lightInfo.color = glm::vec3(0.0f);
+        lightInfo.radius = 1.0f;
+
+        lights.push_back(lightInfo);
+    }
 }
 
 
@@ -85,52 +111,151 @@ void Lab5::Update(float deltaTimeSeconds)
 {
     ClearScreen();
 
-    auto camera = GetSceneCamera();
-
-    // Clear the screen
-    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-
-    // Draw the cubemap
+    for (auto& l : lights)
     {
-        Shader *shader = shaders["ShaderNormal"];
-        shader->Use();
+        // TODO(student): Move the light sources in an orbit around the center of the scene.
+        // The orbit is in the xoz plane. Compute rotationRadians for the current frame 
+        // such that the light sources rotate 6 degrees/second. Use deltaTimeSeconds.
+        float rotationRadians = 0.0f;
 
-        glm::mat4 modelMatrix = glm::scale(glm::mat4(1), glm::vec3(30));
-
-        glUniformMatrix4fv(shader->loc_model_matrix, 1, GL_FALSE, glm::value_ptr(modelMatrix));
-        glUniformMatrix4fv(shader->loc_view_matrix, 1, GL_FALSE, glm::value_ptr(camera->GetViewMatrix()));
-        glUniformMatrix4fv(shader->loc_projection_matrix, 1, GL_FALSE, glm::value_ptr(camera->GetProjectionMatrix()));
-
-        glActiveTexture(GL_TEXTURE0);
-        glBindTexture(GL_TEXTURE_CUBE_MAP, cubeMapTextureID);
-        int loc_texture = shader->GetUniformLocation("texture_cubemap");
-        glUniform1i(loc_texture, 0);
-
-        meshes["cube"]->Render();
+        glm::mat4 rotateMatrix = glm::rotate(glm::mat4(1.0f), rotationRadians, glm::vec3(0, 1, 0));
+        l.position = rotateMatrix * glm::vec4(l.position, 1.0f);
     }
 
-    // Draw the reflection on the mesh
+
+    // ------------------------------------------------------------------------
+    // Deferred rendering pass
     {
-        Shader *shader = shaders["CubeMap"];
+        frameBuffer->Bind();
+
+        auto shader = shaders["Render2Texture"];
+
+        TextureManager::GetTexture("default.png")->BindToTextureUnit(GL_TEXTURE0);
+
+        // Render scene objects
+        RenderMesh(meshes["box"], shader, glm::vec3(1.5, 0.5f, 0), glm::vec3(0.5f));
+        RenderMesh(meshes["box"], shader, glm::vec3(0, 1.05f, 0), glm::vec3(2));
+        RenderMesh(meshes["box"], shader, glm::vec3(-2, 1.5f, 0));
+        RenderMesh(meshes["sphere"], shader, glm::vec3(-4, 1, 1));
+
+        // Render a simple point light bulb for each light (for debugging purposes)
+        TextureManager::GetTexture("default.png")->BindToTextureUnit(GL_TEXTURE0);
+        for (auto &l : lights)
+        {
+            auto model = glm::translate(glm::mat4(1), l.position);
+            model = glm::scale(model, glm::vec3(0.2f));
+            RenderMesh(meshes["sphere"], shader, model);
+        }
+
+        TextureManager::GetTexture("ground.jpg")->BindToTextureUnit(GL_TEXTURE0);
+        RenderMesh(meshes["plane"], shader, glm::vec3(0, 0, 0), glm::vec3(0.5f));
+    }
+
+    // ------------------------------------------------------------------------
+    // Lighting pass
+    {
+        glm::vec3 ambientLight(0.2f);
+        //Set the initial light accumulation in each pixel to be equal to the ambient light.
+        lightBuffer->SetClearColor(glm::vec4(ambientLight.x, ambientLight.y, ambientLight.z, 1.0f));
+        lightBuffer->Bind();
+        glClearColor(0, 0, 0, 1);
+
+        // Enable buffer color accumulation
+        glDepthMask(GL_FALSE);
+        glEnable(GL_BLEND);
+        glBlendEquation(GL_FUNC_ADD);
+        glBlendFunc(GL_ONE, GL_ONE);
+
+        auto shader = shaders["LightPass"];
         shader->Use();
 
-        glm::mat4 modelMatrix = glm::scale(glm::mat4(1), glm::vec3(0.1f));
+        {
+            int texturePositionsLoc = shader->GetUniformLocation("texture_position");
+            glUniform1i(texturePositionsLoc, 0);
+            frameBuffer->BindTexture(0, GL_TEXTURE0);
+        }
 
-        glUniformMatrix4fv(shader->loc_model_matrix, 1, GL_FALSE, glm::value_ptr(modelMatrix));
-        glUniformMatrix4fv(shader->loc_view_matrix, 1, GL_FALSE, glm::value_ptr(camera->GetViewMatrix()));
-        glUniformMatrix4fv(shader->loc_projection_matrix, 1, GL_FALSE, glm::value_ptr(camera->GetProjectionMatrix()));
+        {
+            int textureNormalsLoc = shader->GetUniformLocation("texture_normal");
+            glUniform1i(textureNormalsLoc, 1);
+            frameBuffer->BindTexture(1, GL_TEXTURE0 + 1);
+        }
 
-        auto cameraPosition = camera->m_transform->GetWorldPosition();
+        auto camera = GetSceneCamera();
+        glm::vec3 cameraPos = camera->m_transform->GetWorldPosition();
+        int loc_eyePosition = shader->GetUniformLocation("eye_position");
+        glUniform3fv(loc_eyePosition, 1, glm::value_ptr(cameraPos));
 
-        glActiveTexture(GL_TEXTURE0);
-        glBindTexture(GL_TEXTURE_CUBE_MAP, cubeMapTextureID);
-        int loc_texture = shader->GetUniformLocation("texture_cubemap");
-        glUniform1i(loc_texture, 0);
+        auto resolution = window->GetResolution();
+        int loc_resolution = shader->GetUniformLocation("resolution");
+        glUniform2i(loc_resolution, resolution.x, resolution.y);
 
-        int loc_camera = shader->GetUniformLocation("camera_position");
-        glUniform3f(loc_camera, cameraPosition.x, cameraPosition.y, cameraPosition.z);
+        //Front face culling
+        glEnable(GL_CULL_FACE);
+        glCullFace(GL_FRONT);
 
-        meshes["bunny"]->Render();
+        for (auto& lightInfo : lights)
+        {
+            // TODO(student): Set the shader uniforms 'light_position', 'light_color' and 'light_radius'
+            // with the values from the light source. Use shader 'shader'.
+            
+
+
+            // TODO(student): Draw the mesh "sphere" at the position of the light source
+            // and scaled 2 times the light source radius.
+            // Use RenderMesh(mesh, shader, position, scale). Use shader 'shader'.
+
+        }
+
+        glDisable(GL_CULL_FACE);
+
+        glDepthMask(GL_TRUE);
+        glDisable(GL_BLEND);
+    }
+
+    // ------------------------------------------------------------------------
+    // Composition pass
+    {
+        FrameBuffer::BindDefault();
+
+        auto shader = shaders["Composition"];
+        shader->Use();
+
+        int outputTypeLoc = shader->GetUniformLocation("output_type");
+        glUniform1i(outputTypeLoc, outputType);
+
+        {
+            int texturePositionsLoc = shader->GetUniformLocation("texture_position");
+            glUniform1i(texturePositionsLoc, 1);
+            frameBuffer->BindTexture(0, GL_TEXTURE0 + 1);
+        }
+
+        {
+            int textureNormalsLoc = shader->GetUniformLocation("texture_normal");
+            glUniform1i(textureNormalsLoc, 2);
+            frameBuffer->BindTexture(1, GL_TEXTURE0 + 2);
+        }
+
+        {
+            int textureColorLoc = shader->GetUniformLocation("texture_color");
+            glUniform1i(textureColorLoc, 3);
+            frameBuffer->BindTexture(2, GL_TEXTURE0 + 3);
+        }
+
+        {
+            int textureDepthLoc = shader->GetUniformLocation("texture_depth");
+            glUniform1i(textureDepthLoc, 4);
+            frameBuffer->BindDepthTexture(GL_TEXTURE0 + 4);
+        }
+
+        {
+            int textureLightLoc = shader->GetUniformLocation("texture_light");
+            glUniform1i(textureLightLoc, 5);
+            lightBuffer->BindTexture(0, GL_TEXTURE0 + 5);
+        }
+
+        // Render the object again but with different properties
+        RenderMesh(meshes["quad"], shader, glm::vec3(0, 0, 0));
     }
 }
 
@@ -141,56 +266,19 @@ void Lab5::FrameEnd()
 }
 
 
-unsigned int Lab5::UploadCubeMapTexture(const std::string &pos_x, const std::string &pos_y, const std::string &pos_z, const std::string& neg_x, const std::string& neg_y, const std::string& neg_z)
+void Lab5::LoadShader(const std::string &name)
 {
-    int width, height, chn;
+    std::string shaderPath = PATH_JOIN(window->props.selfDir, SOURCE_PATH::M2, "lab5", "shaders");
 
-    unsigned char* data_pos_x = stbi_load(pos_x.c_str(), &width, &height, &chn, 0);
-    unsigned char* data_pos_y = stbi_load(pos_y.c_str(), &width, &height, &chn, 0);
-    unsigned char* data_pos_z = stbi_load(pos_z.c_str(), &width, &height, &chn, 0);
-    unsigned char* data_neg_x = stbi_load(neg_x.c_str(), &width, &height, &chn, 0);
-    unsigned char* data_neg_y = stbi_load(neg_y.c_str(), &width, &height, &chn, 0);
-    unsigned char* data_neg_z = stbi_load(neg_z.c_str(), &width, &height, &chn, 0);
-
-    unsigned int textureID = 0;
-    // TODO(student): Create the texture
-
-    // TODO(student): Bind the texture
-
-    glEnable(GL_TEXTURE_CUBE_MAP_SEAMLESS);
-
-    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_MIN_FILTER, GL_LINEAR_MIPMAP_LINEAR);
-    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_WRAP_R, GL_CLAMP_TO_EDGE);
-    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_WRAP_S, GL_CLAMP_TO_EDGE);
-    glTexParameteri(GL_TEXTURE_CUBE_MAP, GL_TEXTURE_WRAP_T, GL_CLAMP_TO_EDGE);
-
-    if (GLEW_EXT_texture_filter_anisotropic) {
-        float maxAnisotropy;
-
-        glGetFloatv(GL_MAX_TEXTURE_MAX_ANISOTROPY_EXT, &maxAnisotropy);
-        glTexParameterf(GL_TEXTURE_2D, GL_TEXTURE_MAX_ANISOTROPY_EXT, maxAnisotropy);
-    }
-
-    glPixelStorei(GL_UNPACK_ALIGNMENT, 1);
-
-    // TODO(student): Load texture information for each face
-
-    glGenerateMipmap(GL_TEXTURE_CUBE_MAP);
-    if (GetOpenGLError() == GL_INVALID_OPERATION)
+    // Create a shader program for particle system
     {
-        cout << "\t[NOTE] : For students : DON'T PANIC! This error should go away when completing the tasks." << std::endl;
+        Shader *shader = new Shader(name);
+        shader->AddShader(PATH_JOIN(shaderPath, name + ".VS.glsl"), GL_VERTEX_SHADER);
+        shader->AddShader(PATH_JOIN(shaderPath, name + ".FS.glsl"), GL_FRAGMENT_SHADER);
+
+        shader->CreateAndLink();
+        shaders[shader->GetName()] = shader;
     }
-
-    // Free memory
-    SAFE_FREE(data_pos_x);
-    SAFE_FREE(data_pos_y);
-    SAFE_FREE(data_pos_z);
-    SAFE_FREE(data_neg_x);
-    SAFE_FREE(data_neg_y);
-    SAFE_FREE(data_neg_z);
-
-    return textureID;
 }
 
 
@@ -209,6 +297,15 @@ void Lab5::OnInputUpdate(float deltaTime, int mods)
 void Lab5::OnKeyPress(int key, int mods)
 {
     // Add key press event
+
+    // These are the key mappings for compositing different passes.
+    // What does each key seem to activate? Where can you find the
+    // answer? Examine the source code to find out!
+    int index = key - GLFW_KEY_0;
+    if (index >= 0 && index <= 9)
+    {
+        outputType = index;
+    }
 }
 
 
@@ -245,4 +342,6 @@ void Lab5::OnMouseScroll(int mouseX, int mouseY, int offsetX, int offsetY)
 void Lab5::OnWindowResize(int width, int height)
 {
     // Treat window resize event
+    frameBuffer->Resize(width, height, 32);
+    lightBuffer->Resize(width, height, 32);
 }
