+++ src/lab_m2/lab5/lab5.h
@@ -1,13 +1,22 @@
 #pragma once
 
+#include <string>
+#include <vector>
+
 #include "components/simple_scene.h"
 #include "components/transform.h"
-
-#include <string>
+#include "core/gpu/frame_buffer.h"
 
 
 namespace m2
 {
+    struct LightInfo
+    {
+        glm::vec3 position;
+        glm::vec3 color;
+        float radius;
+    };
+
     class Lab5 : public gfxc::SimpleScene
     {
      public:
@@ -21,8 +30,6 @@ namespace m2
         void Update(float deltaTimeSeconds) override;
         void FrameEnd() override;
 
-        unsigned int UploadCubeMapTexture(const std::string &pos_x, const std::string &pos_y, const std::string &pos_z, const std::string &neg_x, const std::string &neg_y, const std::string &neg_z);
-
         void OnInputUpdate(float deltaTime, int mods) override;
         void OnKeyPress(int key, int mods) override;
         void OnKeyRelease(int key, int mods) override;
@@ -32,7 +39,12 @@ namespace m2
         void OnMouseScroll(int mouseX, int mouseY, int offsetX, int offsetY) override;
         void OnWindowResize(int width, int height) override;
 
+        void LoadShader(const std::string &fileName);
+
      private:
-        int cubeMapTextureID;
+        FrameBuffer *frameBuffer;
+        FrameBuffer *lightBuffer;
+        std::vector<LightInfo> lights;
+        int outputType;
     };
 }   // namespace m2
