+++ src/lab_m2/lab5/shaders/LightPass.VS.glsl
