+++ src/lab_m2/lab5/shaders/Render2Texture.FS.glsl
