+++ src/lab_m2/lab6/shaders/CubeMap.VS.glsl
