+++ src/lab_m2/lab6/shaders/Normal.VS.glsl
